base dapper

+++ backend/src/Volunteers.Application/Models/PagedList.cs
@@ -5,7 +5,7 @@ public class PagedList<T>
     public IReadOnlyList<T> Items { get; init; } = [];
     public int Page { get; init; }
     public int PageSize { get; init; }
-    public int TotalCount { get; init; }
+    public long TotalCount { get; init; }
     public bool IsHasNaxtPage => Page * PageSize < TotalCount;
     public bool IsHasЗreviousPage => Page > 1;
 }
